requirements.txt
python-decouple==3.8
discord.py==2.4.0
PyNaCl==1.5.0
SQLAlchemy~=2.0.36
